--- THIRD_PARTY.md
@@ -1,15 +1,0 @@
-** Portions of this code were adapted from amazon-guardduty-multiaccount-scripts; - https://github.com/aws-samples/amazon-guardduty-multiaccount-scripts
-Copyright (c) 2019 Amazon.com, Inc
-
-Licensed under the Apache License, Version 2.0 (the “License”); you may not use this file except in compliance with the License. You may obtain a copy of the License at
-
-https://www.apache.org/licenses/LICENSE-2.0
-Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an “AS IS” BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
-
-** boto3; version 1.11.9 - https://pypi.org/project/boto3/
-Copyright (c) 2020 Amazon Web Services
-
-Licensed under the Apache License, Version 2.0 (the “License”); you may not use this file except in compliance with the License. You may obtain a copy of the License at
-
-https://www.apache.org/licenses/LICENSE-2.0
-Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an “AS IS” BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
